requirements.txt
flask
load_dotenv # 加载环境变量
gql # GrqphQL 库